LayoutTests/platform/wpe/http/tests/storage/storage-map-leaking-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.
CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.
CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.
CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.
Make sure that StorageAreaMap objects do no leak.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL Unexpected error: QuotaExceededError: The quota has been exceeded.
FAIL successfullyParsed should be true. Was false.
Some tests failed.

TEST COMPLETE
